pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "customer-support-agent"
version = "0.1.0"
description = "A RAG-based AI customer support agent using Azure OpenAI and ChromaDB"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.10"
authors = [
    {name = "Your Name", email = "your.email@example.com"}
]
keywords = ["rag", "chatbot", "azure-openai", "customer-support", "ai-agent"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]

dependencies = [
    "python-dotenv>=1.0.0",
    "requests>=2.31.0",
    "tiktoken>=0.5.0",
    "chromadb>=0.4.22",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
    "black>=23.0.0",
    "isort>=5.12.0",
    "pylint>=3.0.0",
    "mypy>=1.5.0",
    "pre-commit>=3.5.0",
]

[project.urls]
Homepage = "https://github.com/yourusername/customer-support-agent"
Documentation = "https://github.com/yourusername/customer-support-agent#readme"
Repository = "https://github.com/yourusername/customer-support-agent.git"
Issues = "https://github.com/yourusername/customer-support-agent/issues"

[project.scripts]
customer-support-agent = "src.cli:main"

[tool.setuptools.packages.find]
where = ["."]
include = ["src*"]

[tool.black]
line-length = 100
target-version = ['py310', 'py311', 'py312']
include = '\.pyi?$'
exclude = '''
/(
    \.git
    | \.hg
    | \.mypy_cache
    | \.tox
    | \.venv
    | _build
    | buck-out
    | build
    | dist
    | vectorstore
)/
'''

[tool.isort]
profile = "black"
line_length = 100
skip_gitignore = true
skip = ["vectorstore", ".venv", "venv"]

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
ignore_missing_imports = true
exclude = ["vectorstore", "venv", ".venv"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_functions = ["test_*"]
addopts = "-v --tb=short"
filterwarnings = [
    "ignore::DeprecationWarning",
]

[tool.coverage.run]
source = ["src"]
omit = ["tests/*", "**/__init__.py"]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise NotImplementedError",
    "if __name__ == .__main__.:",
    "if TYPE_CHECKING:",
]

[tool.pylint.messages_control]
disable = [
    "C0114",  # missing-module-docstring (handled by other tools)
    "C0115",  # missing-class-docstring (we use them, but not enforced)
    "C0116",  # missing-function-docstring (we use them, but not enforced)
    "R0903",  # too-few-public-methods
    "W0511",  # fixme (allow TODO comments)
]

[tool.pylint.format]
max-line-length = 100

[tool.pylint.basic]
good-names = ["i", "j", "k", "e", "f", "id", "db", "df"]
